refactor: use cobra to implement the dfget client

1. use cobra to implement the dfget client
2. delete useless code of dfget
3. fix some errors about format or lint

Related to: https://github.com/dragonflyoss/Dragonfly/issues/249
